Crate dokany_sys

Structs

BY_HANDLE_FILE_INFORMATION

Required features: "Win32_Storage_FileSystem", "Win32_Foundation"

DOKAN_FILE_INFO

Dokan file information on the current operation.

DOKAN_MOUNT_POINT_INFO

Dokan Mount point information

DOKAN_OPERATIONS

Dokan API callbacks interface

DOKAN_OPTIONS

Dokan mount options used to describe Dokan device behavior.

FILETIME

Required features: "Win32_Foundation"

SECURITY_DESCRIPTOR

Required features: "Win32_Security", "Win32_Foundation"

WIN32_FIND_DATAW

Required features: "Win32_Storage_FileSystem", "Win32_Foundation"

WIN32_FIND_STREAM_DATA

Required features: "Win32_Storage_FileSystem"

Constants

ACCESS_SYSTEM_SECURITY

Required features: "Win32_System_SystemServices"

DELETE

Required features: "Win32_Storage_FileSystem"

DOKAN_DRIVER_INSTALL_ERROR

Dokan mount failed - Can’t install driver.

DOKAN_DRIVE_LETTER_ERROR

Dokan mount failed - Bad drive letter.

DOKAN_ERROR

Dokan mount error.

DOKAN_MOUNT_ERROR

Dokan mount failed.

DOKAN_MOUNT_POINT_ERROR

Dokan mount failed.

DOKAN_OPTION_ALLOW_IPC_BATCHING

Pull batches of events from the driver instead of a single one and execute them parallelly.

DOKAN_OPTION_ALT_STREAM

Enable the use of alternate stream paths in the form :. If this is not specified then the driver will

DOKAN_OPTION_CASE_SENSITIVE

Enable Case sensitive path.

DOKAN_OPTION_CURRENT_SESSION

Mount the drive on current session only

DOKAN_OPTION_DEBUG

Enable ouput debug message

DOKAN_OPTION_DISPATCH_DRIVER_LOGS

Forward the kernel driver global and volume logs to the userland.

DOKAN_OPTION_ENABLE_UNMOUNT_NETWORK_DRIVE

Allows unmounting of network drive via explorer

DOKAN_OPTION_FILELOCK_USER_MODE

Enable Lockfile/Unlockfile operations.

DOKAN_OPTION_MOUNT_MANAGER

Use Windows Mount Manager.

DOKAN_OPTION_NETWORK

Use network drive - Dokan network provider needs to be installed and a DOKAN_OPTIONS.UNCName provided

DOKAN_OPTION_REMOVABLE

Use removable drive

DOKAN_OPTION_STDERR

Enable ouput debug message to stderr

DOKAN_OPTION_WRITE_PROTECT

Enable mount drive as write-protected

DOKAN_START_ERROR

Dokan mount failed - Driver answer that something is wrong.

DOKAN_SUCCESS

Dokan mount succeed.

DOKAN_VERSION_ERROR

Dokan mount failed.

FALSE

Required features: "Win32_Foundation"

FILE_ADD_FILE

Required features: "Win32_Storage_FileSystem"

FILE_ADD_SUBDIRECTORY

Required features: "Win32_Storage_FileSystem"

FILE_ALL_ACCESS

Required features: "Win32_Storage_FileSystem"

FILE_APPEND_DATA

Required features: "Win32_Storage_FileSystem"

FILE_CASE_PRESERVED_NAMES

Required features: "Win32_System_SystemServices"

FILE_CASE_SENSITIVE_SEARCH

Required features: "Win32_System_SystemServices"

FILE_CREATE_PIPE_INSTANCE

Required features: "Win32_Storage_FileSystem"

FILE_DAX_VOLUME

Required features: "Win32_System_SystemServices"

FILE_DELETE_CHILD

Required features: "Win32_Storage_FileSystem"

FILE_EXECUTE

Required features: "Win32_Storage_FileSystem"

FILE_FILE_COMPRESSION

Required features: "Win32_System_SystemServices"

FILE_LIST_DIRECTORY

Required features: "Win32_Storage_FileSystem"

FILE_NAMED_STREAMS

Required features: "Win32_System_SystemServices"

FILE_PERSISTENT_ACLS

Required features: "Win32_System_SystemServices"

FILE_READ_ATTRIBUTES

Required features: "Win32_Storage_FileSystem"

FILE_READ_DATA

Required features: "Win32_Storage_FileSystem"

FILE_READ_EA

Required features: "Win32_Storage_FileSystem"

FILE_READ_ONLY_VOLUME

Required features: "Win32_System_SystemServices"

FILE_RETURNS_CLEANUP_RESULT_INFO

Required features: "Win32_System_SystemServices"

FILE_SEQUENTIAL_WRITE_ONCE

Required features: "Win32_System_SystemServices"

FILE_SUPPORTS_BLOCK_REFCOUNTING

Required features: "Win32_System_SystemServices"

FILE_SUPPORTS_ENCRYPTION

Required features: "Win32_System_SystemServices"

FILE_SUPPORTS_EXTENDED_ATTRIBUTES

Required features: "Win32_System_SystemServices"

FILE_SUPPORTS_GHOSTING

Required features: "Win32_System_SystemServices"

FILE_SUPPORTS_HARD_LINKS

Required features: "Win32_System_SystemServices"

FILE_SUPPORTS_INTEGRITY_STREAMS

Required features: "Win32_System_SystemServices"

FILE_SUPPORTS_OBJECT_IDS

Required features: "Win32_System_SystemServices"

FILE_SUPPORTS_OPEN_BY_FILE_ID

Required features: "Win32_System_SystemServices"

FILE_SUPPORTS_POSIX_UNLINK_RENAME

Required features: "Win32_System_SystemServices"

FILE_SUPPORTS_REMOTE_STORAGE

Required features: "Win32_System_SystemServices"

FILE_SUPPORTS_REPARSE_POINTS

Required features: "Win32_System_SystemServices"

FILE_SUPPORTS_SPARSE_FILES

Required features: "Win32_System_SystemServices"

FILE_SUPPORTS_SPARSE_VDL

Required features: "Win32_System_SystemServices"

FILE_SUPPORTS_TRANSACTIONS

Required features: "Win32_System_SystemServices"

FILE_SUPPORTS_USN_JOURNAL

Required features: "Win32_System_SystemServices"

FILE_TRAVERSE

Required features: "Win32_Storage_FileSystem"

FILE_UNICODE_ON_DISK

Required features: "Win32_System_SystemServices"

FILE_VOLUME_IS_COMPRESSED

Required features: "Win32_System_SystemServices"

FILE_VOLUME_QUOTAS

Required features: "Win32_System_SystemServices"

FILE_WRITE_ATTRIBUTES

Required features: "Win32_Storage_FileSystem"

FILE_WRITE_DATA

Required features: "Win32_Storage_FileSystem"

FILE_WRITE_EA

Required features: "Win32_Storage_FileSystem"

GENERIC_ALL

Required features: "Win32_System_SystemServices"

GENERIC_EXECUTE

Required features: "Win32_System_SystemServices"

GENERIC_READ

Required features: "Win32_System_SystemServices"

GENERIC_WRITE

Required features: "Win32_System_SystemServices"

MAXIMUM_ALLOWED

Required features: "Win32_System_SystemServices"

MAX_PATH

Required features: "Win32_Foundation"

READ_CONTROL

Required features: "Win32_Storage_FileSystem"

SPECIFIC_RIGHTS_ALL

Required features: "Win32_Storage_FileSystem"

STANDARD_RIGHTS_ALL

Required features: "Win32_Storage_FileSystem"

STANDARD_RIGHTS_EXECUTE

Required features: "Win32_Storage_FileSystem"

STANDARD_RIGHTS_READ

Required features: "Win32_Storage_FileSystem"

STANDARD_RIGHTS_REQUIRED

Required features: "Win32_Storage_FileSystem"

STANDARD_RIGHTS_WRITE

Required features: "Win32_Storage_FileSystem"

STATUS_INTERNAL_ERROR

Required features: "Win32_Foundation"

STATUS_NOT_IMPLEMENTED

Required features: "Win32_Foundation"

STATUS_NO_SUCH_FILE

Required features: "Win32_Foundation"

STATUS_SUCCESS

Required features: "Win32_Foundation"

SYNCHRONIZE

Required features: "Win32_Storage_FileSystem"

TRUE

Required features: "Win32_Foundation"

VOLUME_SECURITY_DESCRIPTOR_MAX_SIZE

This is arbitrary. There isn’t really an absolute max, but we marshal it in a fixed-size buffer.

WRITE_DAC

Required features: "Win32_Storage_FileSystem"

WRITE_OWNER

Required features: "Win32_Storage_FileSystem"

Functions

DokanCloseHandle⚠

Unmount the Dokan instance.

DokanCreateFileSystem⚠

Mount a new Dokan Volume.

DokanDriverVersion⚠

Get the version of the Dokan driver.

DokanGetMountPointList⚠

Get active Dokan mount points.

DokanInit⚠

Initialize all required Dokan internal resources.

DokanIsFileSystemRunning⚠

Check if the FileSystem is still running or not.

DokanIsNameInExpression⚠

Checks whether Name matches Expression

DokanMain⚠

Mount a new Dokan Volume.

DokanOpenRequestorToken⚠

Get the handle to Access Token.

DokanReleaseMountPointList⚠

Release Mount point list resources from DokanGetMountPointList.

DokanRemoveMountPoint⚠

Unmount a Dokan device from a mount point

DokanResetTimeout⚠

Extends the timeout of the current IO operation in driver.

DokanShutdown⚠

Release all allocated resources by DokanInit when they are no longer needed.

DokanUnmount⚠

Unmount a Dokan device from a driver letter.

DokanVersion⚠

Get the version of Dokan.

DokanWaitForFileSystemClosed⚠

Wait until the FileSystem is unmount.

Type Definitions

ACCESS_MASK

BOOL

BOOLEAN

CHAR

CleanupCallback

CloseFileCallback

DOKAN_HANDLE

DWORD

DeleteDirectoryCallback

DeleteFileCallback

DokanMainResult

DokanOptionFlag

FILE_ACCESS_FLAGS

Required features: "Win32_Storage_FileSystem"

FindFilesCallback

FindFilesWithPatternCallback

FindStreamsCallback

FlushFileBuffersCallback

GetDiskFreeSpaceCallback

GetFileInformationCallback

GetFileSecurityCallback

GetVolumeInformationCallback

HANDLE

LONGLONG

LPBY_HANDLE_FILE_INFORMATION

LPCVOID

LPCWSTR

LPDWORD

LPVOID

LPWSTR

LockFileCallback

MountedCallback

MoveFileCallback

NTSTATUS

PCWSTR

PDOKAN_FILE_INFO

PDOKAN_IO_SECURITY_CONTEXT

PDOKAN_MOUNT_POINT_INFO

PDOKAN_OPERATIONS

PDOKAN_OPTIONS

PFillFindData

FillFindData Used to add an entry in FindFiles operation

PFillFindStreamData

FillFindStreamData Used to add an entry in FindStreams

PSECURITY_DESCRIPTOR

PSECURITY_INFORMATION

PULONG

PULONGLONG

PVOID

PWIN32_FIND_DATAW

PWIN32_FIND_STREAM_DATA

PWSTR

ReadFileCallback

SECURITY_INFORMATION

SetAllocationSizeCallback

SetEndOfFileCallback

SetFileAttributesCallback

SetFileSecurityCallback

SetFileTimeCallback

UCHAR

ULONG

ULONG64

ULONGLONG

USHORT

UnlockFileCallback

Unmounted

WCHAR

WriteFileCallback

ZwCreateFileCallback

c_int

Equivalent to C’s signed int (int) type.

c_void

Equivalent to C’s void type when used as a pointer.