Struct dokany_sys::DOKAN_OPTIONS

#[repr(C)]
pub struct DOKAN_OPTIONS {
    pub Version: USHORT,
    pub SingleThread: BOOLEAN,
    pub Options: ULONG,
    pub GlobalContext: ULONG64,
    pub MountPoint: LPCWSTR,
    pub UNCName: LPCWSTR,
    pub Timeout: ULONG,
    pub AllocationUnitSize: ULONG,
    pub SectorSize: ULONG,
    pub VolumeSecurityDescriptorLength: ULONG,
    pub VolumeSecurityDescriptor: [CHAR; 16384],
}

Dokan mount options used to describe Dokan device behavior.

Fields
Version: USHORT

Version of the Dokan features requested without dots (version “123” is equal to Dokan version 1.2.3).

SingleThread: BOOLEAN

Only use a single thread to process events. This is highly not recommended as can easily create a bottleneck.

Options: ULONG

Features enabled for the mount.

GlobalContext: ULONG64

FileSystem can store anything here.

MountPoint: LPCWSTR

Mount point. It can be a driver letter like “M:" or a folder path “C:\mount\dokan” on a NTFS partition.

UNCName: LPCWSTR

UNC Name for the Network Redirector see Support for UNC Naming

Timeout: ULONG

Max timeout in milliseconds of each request before Dokan gives up to wait events to complete. A timeout request is a sign that the userland implementation is no longer able to properly manage requests in time. The driver will therefore unmount the device when a timeout trigger in order to keep the system stable. The default timeout value is 15 seconds.

AllocationUnitSize: ULONG

Allocation Unit Size of the volume. This will affect the file size.

SectorSize: ULONG

Sector Size of the volume. This will affect the file size.

VolumeSecurityDescriptorLength: ULONG

Length of the optional VolumeSecurityDescriptor provided. Set 0 will disable the option.

VolumeSecurityDescriptor: [CHAR; 16384]

Optional Volume Security descriptor.

See InitializeSecurityDescriptor

Implementations

impl DOKAN_OPTIONS

pub fn new() -> Self

Make an empty DOKAN_OPTIONS.